Remove usages of 'window.readerControl'
