+++ src/apis/setMaxZoomLevel.js
@@ -9,20 +9,6 @@ WebViewer(...)
   });
  */
 
-import actions from 'actions';
-import zoomFactors, { defaultZoomList } from 'constants/zoomFactors';
-import getActualZoomLevel from 'helpers/getActualZoomLevel';
+import setMaxZoom from 'helpers/setMaxZoom';
 
-export default store => zoomLevel => {
-  const maxZoom = getActualZoomLevel(zoomLevel);
-
-  if (maxZoom) {
-    const minZoom = zoomFactors.getMinZoomLevel();
-    const zoomList = defaultZoomList.filter(zoom => zoom <= maxZoom && zoom >= minZoom);
-    zoomFactors.setMaxZoomLevel(maxZoom);
-    store.dispatch(actions.setZoomList(zoomList));
-    window.Tools.MarqueeZoomTool.setMaxZoomLevel(maxZoom);
-  } else {
-    console.warn('Type of the argument for setMaxZoomLevel must be either string or number');
-  }
-};
+export default store => setMaxZoom(store.dispatch);
