+++ src/components/ReaderModeViewer/ReaderModeViewer.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import core from 'core';
 import zoomFactors from 'constants/zoomFactors';
+import { connect } from 'react-redux';
+import setMaxZoom from 'helpers/setMaxZoom';
 
 class ReaderModeViewer extends React.PureComponent {
   constructor(props) {
@@ -27,7 +29,7 @@ class ReaderModeViewer extends React.PureComponent {
     core.removeEventListener('pageNumberUpdated', this.goToPage);
     core.removeEventListener('zoomUpdated', this.setZoom);
 
-    window.readerControl.setMaxZoomLevel(this.originalMaxZoom);
+    setMaxZoom(this.props.dispatch)(this.originalMaxZoom);
   }
 
   componentDidUpdate(prevProps) {
@@ -77,12 +79,12 @@ class ReaderModeViewer extends React.PureComponent {
   updateMaxZoom() {
     // Calling the FitWidth function to get the calculated fit width zoom level for normal page rendering
     const maxZoomLevel = core.getDocumentViewer().FitMode.FitWidth.call(core.getDocumentViewer());
-    window.readerControl.setMaxZoomLevel(maxZoomLevel);
+    setMaxZoom(this.props.dispatch)(maxZoomLevel);
 
     if (maxZoomLevel < core.getZoom()) {
-      window.readerControl.setZoomLevel(maxZoomLevel);
+      core.zoomTo(maxZoomLevel);
     }
   }
 }
 
-export default ReaderModeViewer;
+export default connect()(ReaderModeViewer);
